move minesweeper questions into org.batfish.minesweeper.question
