+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/question/SmtBlackholeQuestionPlugin.java
@@ -1,4 +1,4 @@
-package org.batfish.question;
+package org.batfish.minesweeper.question;
 
 import com.google.auto.service.AutoService;
 import org.batfish.common.Answerer;
@@ -9,44 +9,42 @@ import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.questions.Question;
 import org.batfish.datamodel.questions.smt.HeaderQuestion;
 import org.batfish.minesweeper.smt.PropertyChecker;
+import org.batfish.question.QuestionPlugin;
 
 @AutoService(Plugin.class)
-public class SmtForwardingQuestionPlugin extends QuestionPlugin {
+public class SmtBlackholeQuestionPlugin extends QuestionPlugin {
 
-  public static class ForwardingAnswerer extends Answerer {
+  public static class BlackholeAnswerer extends Answerer {
 
-    public ForwardingAnswerer(Question question, IBatfish batfish) {
+    public BlackholeAnswerer(Question question, IBatfish batfish) {
       super(question, batfish);
     }
 
     @Override
     public AnswerElement answer() {
-      ForwardingQuestion q = (ForwardingQuestion) _question;
+      BlackholeQuestion q = (BlackholeQuestion) _question;
       PropertyChecker p = new PropertyChecker(new BDDPacket(), _batfish);
-      return p.checkForwarding(q);
+      return p.checkBlackHole(q);
     }
   }
 
-  public static class ForwardingQuestion extends HeaderQuestion {
+  public static class BlackholeQuestion extends HeaderQuestion {
 
-    @Override
-    public boolean getDataPlane() {
-      return false;
-    }
+    public BlackholeQuestion() {}
 
     @Override
     public String getName() {
-      return "smt-forwarding";
+      return "smt-blackhole";
     }
   }
 
   @Override
   protected Answerer createAnswerer(Question question, IBatfish batfish) {
-    return new ForwardingAnswerer(question, batfish);
+    return new BlackholeAnswerer(question, batfish);
   }
 
   @Override
   protected Question createQuestion() {
-    return new ForwardingQuestion();
+    return new BlackholeQuestion();
   }
 }
