+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/question/SmtForwardingQuestionPlugin.java
@@ -1,4 +1,4 @@
-package org.batfish.question;
+package org.batfish.minesweeper.question;
 
 import com.google.auto.service.AutoService;
 import org.batfish.common.Answerer;
@@ -7,26 +7,28 @@ import org.batfish.common.plugin.IBatfish;
 import org.batfish.common.plugin.Plugin;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.questions.Question;
-import org.batfish.datamodel.questions.smt.HeaderLocationQuestion;
+import org.batfish.datamodel.questions.smt.HeaderQuestion;
 import org.batfish.minesweeper.smt.PropertyChecker;
+import org.batfish.question.QuestionPlugin;
 
 @AutoService(Plugin.class)
-public class SmtEqualLengthQuestionPlugin extends QuestionPlugin {
+public class SmtForwardingQuestionPlugin extends QuestionPlugin {
 
-  public static class EqualLengthAnswerer extends Answerer {
+  public static class ForwardingAnswerer extends Answerer {
 
-    public EqualLengthAnswerer(Question question, IBatfish batfish) {
+    public ForwardingAnswerer(Question question, IBatfish batfish) {
       super(question, batfish);
     }
 
     @Override
     public AnswerElement answer() {
+      ForwardingQuestion q = (ForwardingQuestion) _question;
       PropertyChecker p = new PropertyChecker(new BDDPacket(), _batfish);
-      return p.checkEqualLength((EqualLengthQuestion) _question);
+      return p.checkForwarding(q);
     }
   }
 
-  public static class EqualLengthQuestion extends HeaderLocationQuestion {
+  public static class ForwardingQuestion extends HeaderQuestion {
 
     @Override
     public boolean getDataPlane() {
@@ -35,17 +37,17 @@ public class SmtEqualLengthQuestionPlugin extends QuestionPlugin {
 
     @Override
     public String getName() {
-      return "smt-equal-length";
+      return "smt-forwarding";
     }
   }
 
   @Override
   protected Answerer createAnswerer(Question question, IBatfish batfish) {
-    return new EqualLengthAnswerer(question, batfish);
+    return new ForwardingAnswerer(question, batfish);
   }
 
   @Override
   protected Question createQuestion() {
-    return new EqualLengthQuestion();
+    return new ForwardingQuestion();
   }
 }
