+++ projects/minesweeper/src/main/java/org/batfish/minesweeper/question/SmtMultipathConsistencyQuestionPlugin.java
@@ -1,4 +1,4 @@
-package org.batfish.question;
+package org.batfish.minesweeper.question;
 
 import com.google.auto.service.AutoService;
 import org.batfish.common.Answerer;
@@ -7,27 +7,28 @@ import org.batfish.common.plugin.IBatfish;
 import org.batfish.common.plugin.Plugin;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.questions.Question;
-import org.batfish.datamodel.questions.smt.HeaderQuestion;
+import org.batfish.datamodel.questions.smt.HeaderLocationQuestion;
 import org.batfish.minesweeper.smt.PropertyChecker;
+import org.batfish.question.QuestionPlugin;
 
 @AutoService(Plugin.class)
-public class SmtRoutingLoopQuestionPlugin extends QuestionPlugin {
+public class SmtMultipathConsistencyQuestionPlugin extends QuestionPlugin {
 
-  public static class RoutingLoopAnswerer extends Answerer {
+  public static class MulipathConsistencyAnswerer extends Answerer {
 
-    public RoutingLoopAnswerer(Question question, IBatfish batfish) {
+    public MulipathConsistencyAnswerer(Question question, IBatfish batfish) {
       super(question, batfish);
     }
 
     @Override
     public AnswerElement answer() {
-      RoutingLoopQuestion q = (RoutingLoopQuestion) _question;
+      MultipathConsistencyQuestion q = (MultipathConsistencyQuestion) _question;
       PropertyChecker p = new PropertyChecker(new BDDPacket(), _batfish);
-      return p.checkRoutingLoop(q);
+      return p.checkMultipathConsistency(q);
     }
   }
 
-  public static class RoutingLoopQuestion extends HeaderQuestion {
+  public static class MultipathConsistencyQuestion extends HeaderLocationQuestion {
 
     @Override
     public boolean getDataPlane() {
@@ -36,17 +37,17 @@ public class SmtRoutingLoopQuestionPlugin extends QuestionPlugin {
 
     @Override
     public String getName() {
-      return "smt-routing-loop";
+      return "smt-multipath-consistency";
     }
   }
 
   @Override
   protected Answerer createAnswerer(Question question, IBatfish batfish) {
-    return new RoutingLoopAnswerer(question, batfish);
+    return new MulipathConsistencyAnswerer(question, batfish);
   }
 
   @Override
   protected Question createQuestion() {
-    return new RoutingLoopQuestion();
+    return new MultipathConsistencyQuestion();
   }
 }
